ARQ-219 Add Spring Embedded Container support
Spring Embedded Container allow deployments of type JAR and WAR archives.
Currently it may be used for testing buissnes objects. It does not mock the
servlet container so it's not capable of testing web applications.

The embedded container uses the ApplicationProducer from inetgration module in
order to instatiate the Spring's application context used during tests.